backend/requirements.txt
txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
pydantic
email-validator
python-multipart
